@@ -17,10 +17,10 @@
         "mergely": "^4.3.9",
         "moment": "^2.29.4",
         "qs": "^6.11.1",
-        "react": "^16.14.0",
+        "react": "^18.2.0",
         "react-app-rewired": "^2.1.8",
         "react-day-picker": "^7.4.10",
-        "react-dom": "^16.14.0",
+        "react-dom": "^18.2.0",
         "react-grid-layout": "^0.16.6",
         "react-highcharts": "^16.1.0",
         "react-highlight-words": "^0.16.0",
